+++ django_filters/rest_framework/filterset.py
@@ -3,11 +3,12 @@ from __future__ import absolute_import
 from copy import deepcopy
 
 from django.db import models
+from django import forms
 from django.utils.translation import ugettext_lazy as _
 
 from django_filters import filterset
 from .filters import BooleanFilter, IsoDateTimeFilter
-from .. import compat
+from .. import compat, utils
 
 
 FILTER_FOR_DBFIELD_DEFAULTS = deepcopy(filterset.FILTER_FOR_DBFIELD_DEFAULTS)
@@ -20,14 +21,15 @@ FILTER_FOR_DBFIELD_DEFAULTS.update({
 class FilterSet(filterset.FilterSet):
     FILTER_DEFAULTS = FILTER_FOR_DBFIELD_DEFAULTS
 
-    def __init__(self, *args, **kwargs):
-        super(FilterSet, self).__init__(*args, **kwargs)
+    @property
+    def form(self):
+        form = super(FilterSet, self).form
 
         if compat.is_crispy():
             from crispy_forms.helper import FormHelper
             from crispy_forms.layout import Layout, Submit
 
-            layout_components = list(self.form.fields.keys()) + [
+            layout_components = list(form.fields.keys()) + [
                 Submit('', _('Submit'), css_class='btn-default'),
             ]
             helper = FormHelper()
@@ -35,4 +37,15 @@ class FilterSet(filterset.FilterSet):
             helper.template_pack = 'bootstrap3'
             helper.layout = Layout(*layout_components)
 
-            self.form.helper = helper
+            form.helper = helper
+
+        return form
+
+    @property
+    def qs(self):
+        from rest_framework.exceptions import ValidationError
+
+        try:
+            return super(FilterSet, self).qs
+        except forms.ValidationError as e:
+            raise ValidationError(utils.raw_validation(e))