+++ CHANGES.rst
@@ -1,3 +1,14 @@
+Version 1.0.3 (2107-05-16)
+--------------------------
+
+Improves compatibility with Django REST Framework schema generation. 
+
+See the `1.0.2 Milestone`__ for full details.
+
+__ https://github.com/carltongibson/django-filter/milestone/13?closed=1
+
+
+
 Version 1.0.2 (2017-03-20)
 --------------------------